[ADAPTEDS-103] Saving progress, likely going to need to build a barebones front end in order to test anything with websocket.
