+++ authorization/src/main/java/com/terabite/chat/WebSocketConfig.java
@@ -8,7 +8,6 @@ import org.springframework.messaging.converter.DefaultContentTypeResolver;
 import org.springframework.messaging.converter.MappingJackson2MessageConverter;
 import org.springframework.messaging.converter.MessageConverter;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.util.MimeType;
 import org.springframework.util.MimeTypeUtils;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
 import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
@@ -21,9 +20,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 public class WebSocketConfig implements WebSocketMessageBrokerConfigurer{
     @Override
     public void configureMessageBroker(MessageBrokerRegistry registry){
-        registry.enableSimpleBroker("/user");
+        //                                                  /user
+        registry.enableSimpleBroker("/chat/user");
         registry.setApplicationDestinationPrefixes("/app");
-        registry.setUserDestinationPrefix("/user");
+        //                                                   /user
+        registry.setUserDestinationPrefix("/chat/user");
     }
 
     @Override
@@ -31,15 +32,15 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer{
         registry.addEndpoint("/ws").withSockJS();
     }
 
-    @Override
-    public boolean configureMessageConverters(List<MessageConverter> messageConverters){
-        DefaultContentTypeResolver resolver=new DefaultContentTypeResolver();
-        resolver.setDefaultMimeType(MimeTypeUtils.APPLICATION_JSON);
-        MappingJackson2MessageConverter converter = new MappingJackson2MessageConverter();
-        converter.setObjectMapper(new ObjectMapper());
-        converter.setContentTypeResolver(resolver);
-        messageConverters.add(converter);
+    // @Override
+    // public boolean configureMessageConverters(List<MessageConverter> messageConverters){
+    //     DefaultContentTypeResolver resolver=new DefaultContentTypeResolver();
+    //     resolver.setDefaultMimeType(MimeTypeUtils.APPLICATION_JSON);
+    //     MappingJackson2MessageConverter converter = new MappingJackson2MessageConverter();
+    //     converter.setObjectMapper(new ObjectMapper());
+    //     converter.setContentTypeResolver(resolver);
+    //     messageConverters.add(converter);
 
-        return false;
-    }
+    //     return false;
+    // }
 }
