+++ authorization/src/main/java/com/terabite/chat/controller/ChatController.java
@@ -1,10 +1,12 @@
 package com.terabite.chat.controller;
 
+import com.terabite.chat.model.ChatUser;
 import com.terabite.chat.model.Message;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
+import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,6 +14,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.terabite.chat.service.ChatNotification;
+import com.terabite.chat.service.ChatUserService;
 import com.terabite.chat.service.MessageService;
 
 @Controller
@@ -19,10 +22,12 @@ import com.terabite.chat.service.MessageService;
 public class ChatController {
     private final SimpMessagingTemplate messagingTemplate;
     private final MessageService messageService;
+    private final ChatUserService chatUserService;
     
-    public ChatController(MessageService messageService, SimpMessagingTemplate messagingTemplate){
+    public ChatController(MessageService messageService, SimpMessagingTemplate messagingTemplate, ChatUserService chatUserService){
         this.messageService=messageService;
         this.messagingTemplate=messagingTemplate;
+        this.chatUserService=chatUserService;
     }
 
     @MessageMapping("/process_message")
@@ -36,5 +41,19 @@ public class ChatController {
     @GetMapping("/messages/{senderId}/{recipientId}")
     public ResponseEntity<List<Message>> findChatMessages(@PathVariable("senderId") String senderId, @PathVariable("recipientId") String recipientId){
         return ResponseEntity.ok(messageService.findChatMessages(senderId, recipientId));
+    }
+
+    @MessageMapping("/chatUser.addUser")
+    //this is a queue that the front end will need to subscribe to
+    @SendTo("/chatUser/topic")
+    //this method will likely need to be called when a user is created in userInformation
+    public ChatUser addChatUser(@Payload ChatUser chatUser){
+        chatUserService.saveChatUser(chatUser);
+        return chatUser;
+    }
+
+    @GetMapping("/chatUsers")
+    public ResponseEntity<List<ChatUser>> findChatUsers(ChatUser chatUser) {
+        return ResponseEntity.ok(chatUserService.findUsers(chatUser));
     }
 }
