+++ authorization/src/main/java/com/terabite/chat/model/Message.java
@@ -1,8 +1,7 @@
 package com.terabite.chat.model;
 
 import jakarta.persistence.Entity;
-//import jakarta.persistence.Id;
-import org.springframework.data.annotation.Id;
+import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import java.util.Date;
 @Entity
